added some remove functions
added implmentation for remove_atFront & remove_atBack.
Something I noticed while writing this block of code was having a small understanding of what is going behind the scenes. On line 76 and 88, if we comment the lines out of the problem, the complier will run forever. Im assuming because the after doing the deleting, the new head/tail node will still point to the old node that we just previously deleted.